Fixed AppDrawer BG Blur
- Fixed and Optimized AppDrawer BG Blur
