+++ res/layout-port/launcher.xml
@@ -15,70 +15,82 @@
 -->
 
 <!-- Full screen view projects under the status bar and contains the background -->
-<com.android.launcher3.LauncherRootView
+<RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:launcher="http://schemas.android.com/apk/res-auto"
-
-    android:id="@+id/launcher"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:fitsSystemWindows="true">
+    android:layout_height="match_parent">
 
-    <com.android.launcher3.dragndrop.DragLayer
-        android:id="@+id/drag_layer"
-        android:clipChildren="false"
-        android:importantForAccessibility="no"
-        android:clipToPadding="false"
-        android:background="@drawable/workspace_bg"
+    <com.android.launcher3.pixperia.CropImageView
+        android:id="@+id/blur_bg"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:scaleType="centerCrop"
+        android:focusable="false" />
 
-        <!-- The workspace contains 5 screens of cells -->
-        <!-- DO NOT CHANGE THE ID -->
-        <com.android.launcher3.Workspace
-            android:id="@+id/workspace"
+    <com.android.launcher3.LauncherRootView
+        android:id="@+id/launcher"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:fitsSystemWindows="true">
+
+        <com.android.launcher3.dragndrop.DragLayer
+            android:id="@+id/drag_layer"
+            android:clipChildren="false"
+            android:importantForAccessibility="no"
+            android:clipToPadding="false"
+            android:background="@drawable/workspace_bg"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            launcher:pageIndicator="@+id/page_indicator">
-        </com.android.launcher3.Workspace>
+            android:layout_height="match_parent">
 
-        <!-- DO NOT CHANGE THE ID -->
-        <include layout="@layout/hotseat"
-            android:id="@+id/hotseat"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            launcher:layout_ignoreInsets="true" />
+            <!-- The workspace contains 5 screens of cells -->
+            <!-- DO NOT CHANGE THE ID -->
+            <com.android.launcher3.Workspace
+                android:id="@+id/workspace"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_gravity="center"
+                launcher:pageIndicator="@+id/page_indicator">
+            </com.android.launcher3.Workspace>
 
-        <include layout="@layout/overview_panel"
-            android:id="@+id/overview_panel"
-            android:visibility="gone" />
+            <!-- DO NOT CHANGE THE ID -->
+            <include layout="@layout/hotseat"
+                android:id="@+id/hotseat"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                launcher:layout_ignoreInsets="true" />
 
-        <!-- Keep these behind the workspace so that they are not visible when
-             we go into AllApps -->
-        <include layout="@layout/page_indicator"
-            android:id="@+id/page_indicator" />
+            <include layout="@layout/overview_panel"
+                android:id="@+id/overview_panel"
+                android:visibility="gone" />
 
-        <include
-            android:id="@+id/drop_target_bar"
-            layout="@layout/drop_target_bar_horz" />
+            <!-- Keep these behind the workspace so that they are not visible when
+                 we go into AllApps -->
+            <include layout="@layout/page_indicator"
+                android:id="@+id/page_indicator" />
 
-        <include
-            layout="@layout/qsb_container"
-            android:visibility="gone"
-            android:id="@+id/qsb_container" />
+            <include
+                android:id="@+id/drop_target_bar"
+                layout="@layout/drop_target_bar_horz" />
 
-        <include layout="@layout/widgets_view"
-            android:id="@+id/widgets_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:visibility="invisible" />
+            <include
+                layout="@layout/qsb_container"
+                android:visibility="gone"
+                android:id="@+id/qsb_container" />
 
-        <include layout="@layout/all_apps"
-            android:id="@+id/apps_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:visibility="invisible" />
-    </com.android.launcher3.dragndrop.DragLayer>
+            <include layout="@layout/widgets_view"
+                android:id="@+id/widgets_view"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:visibility="invisible" />
 
-</com.android.launcher3.LauncherRootView>
+            <include layout="@layout/all_apps"
+                android:id="@+id/apps_view"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:visibility="invisible" />
+        </com.android.launcher3.dragndrop.DragLayer>
+
+
+    </com.android.launcher3.LauncherRootView>
+</RelativeLayout>
